+++ resolwe/flow/models/storage.py
@@ -5,6 +5,35 @@ from django.contrib.postgres.fields import JSONField
 from django.db import models
 
 from .base import BaseModel
+from .functions import JsonGetPath
+from .utils import json_path_components
+
+
+class StorageManager(models.Manager):
+    """Manager for Storage objects."""
+
+    def with_json_path(self, path, field=None):
+        """Annotate Storage objects with a specific JSON path.
+
+        :param path: Path to get inside the stored object, which can be
+            either a list of path components or a comma-separated
+            string
+        :param field: Optional output field name
+        """
+        if field is None:
+            field = '_'.join(['json'] + json_path_components(path))
+
+        kwargs = {field: JsonGetPath('json', path)}
+        return self.defer('json').annotate(**kwargs)
+
+    def get_json_path(self, path):
+        """Return only a specific JSON path of Storage objects.
+
+        :param path: Path to get inside the stored object, which can be
+            either a list of path components or a comma-separated
+            string
+        """
+        return self.with_json_path(path, field='result').values_list('result', flat=True)
 
 
 class Storage(BaseModel):
@@ -15,6 +44,9 @@ class Storage(BaseModel):
 
     #: actual JSON stored
     json = JSONField()
+
+    #: storage manager
+    objects = StorageManager()
 
 
 class LazyStorageJSON(object):